+++ test/test_h5.py
@@ -11,11 +11,11 @@ class TestReadWrite(unittest.TestCase):
     """Write a dataset with multiples views, and load it back. Compare the loaded dataset vs expetations.
     """
     def setUp(self) -> None:
+        """"This will automatically call for EVERY single test we run."""
         self.test_dir = "./test/test_files/"
-        self.fname = self.test_dir + "test_ex1_t2_ch2_illum2_angle2.h5"
         if not os.path.exists(self.test_dir):
             os.mkdir(self.test_dir)
-
+        self.fname = self.test_dir + "test_real_stack.h5"
         self.NZ, self.NY, self.NX = 8, 35, 35 # XY dims must be odd to get nominal 65535 peak value.
         self.N_T, self.N_CH, self.N_ILL, self.N_TILES, self.N_ANGLES = 2, 2, 4, 6, 4
         
@@ -27,7 +27,7 @@ class TestReadWrite(unittest.TestCase):
                                        nchannels=self.N_CH,
                                        nilluminations=self.N_ILL,
                                        ntiles=self.N_TILES,
-                                       nangles=self.N_ANGLES, overwrite=True)
+                                       nangles=self.N_ANGLES)
         for t in range(self.N_T):
             for i_ch in range(self.N_CH):
                 for i_illum in range(self.N_ILL):
@@ -150,6 +150,109 @@ class TestReadWrite(unittest.TestCase):
         editor.finalize()
 
     def tearDown(self) -> None:
+        """"Tidies up after EACH test method has been run."""
+        if os.path.exists(self.test_dir):
+            shutil.rmtree(self.test_dir)
+
+
+class TestReadWriteVirtual(unittest.TestCase):
+    """Write virtual stacks. """
+    def setUp(self) -> None:
+        """"This will automatically call for EVERY single test we run."""
+        self.test_dir = "./test/test_files/"
+        if not os.path.exists(self.test_dir):
+            os.mkdir(self.test_dir)
+        self.fname0 = self.test_dir + "test_virtual_by_plane.h5"
+        self.fname1 = self.test_dir + "test_virtual_by_substack.h5"
+
+        self.NZ, self.NY, self.NX = 8, 256, 256  # XY dims must be odd to get nominal 65535 peak value.
+        self.N_T, self.N_CH, self.N_ILL, self.N_TILES, self.N_ANGLES = 2, 2, 2, 2, 2
+        self.N_SUBSTACKS = 4
+
+        self.plane = np.empty((self.NY, self.NX), "uint16")
+        self.substack = np.empty((self.NZ//self.N_SUBSTACKS, self.NY, self.NX), "uint16")
+        self.plane = generate_test_image((self.NY, self.NX), 1, 2)
+        for z in range(self.NZ//self.N_SUBSTACKS):
+            self.substack[z, :, :] = generate_test_image((self.NY, self.NX), z, self.NZ//self.N_SUBSTACKS)
+
+        self.write_virtual_by_plane()
+        self.write_virtual_by_substack()
+
+    def write_virtual_by_plane(self):
+        bdv_writer = npy2bdv.BdvWriter(self.fname0,
+                                       nchannels=self.N_CH,
+                                       nilluminations=self.N_ILL,
+                                       ntiles=self.N_TILES,
+                                       nangles=self.N_ANGLES)
+        for t in range(self.N_T):
+            for i_ch in range(self.N_CH):
+                for i_illum in range(self.N_ILL):
+                    for i_tile in range(self.N_TILES):
+                        for i_angle in range(self.N_ANGLES):
+                            bdv_writer.append_view(stack=None,
+                                                   virtual_stack_dim=(self.NZ, self.NY, self.NX),
+                                                   time=t,
+                                                   channel=i_ch,
+                                                   illumination=i_illum,
+                                                   tile=i_tile,
+                                                   angle=i_angle)
+                            for iz in range(self.NZ):
+                                bdv_writer.append_plane(plane=self.plane,
+                                                        z=iz,
+                                                        time=t, channel=i_ch, illumination=i_illum,
+                                                        tile=i_tile, angle=i_angle)
+        bdv_writer.write_xml_file(ntimes=self.N_T)
+        bdv_writer.close()
+
+    def write_virtual_by_substack(self):
+        ''''Write a virtual stack by substacks, with downsampling pyramid.'''
+        bdv_writer = npy2bdv.BdvWriter(self.fname1,
+                                       nchannels=self.N_CH,
+                                       nilluminations=self.N_ILL,
+                                       ntiles=self.N_TILES,
+                                       nangles=self.N_ANGLES,
+                                       subsamp=((1, 1, 1), (2, 4, 4)),
+                                       blockdim=((4, 16, 16), ))
+        # Initialize virtual stacks
+        for t in range(self.N_T):
+            for i_ch in range(self.N_CH):
+                for i_illum in range(self.N_ILL):
+                    for i_tile in range(self.N_TILES):
+                        for i_angle in range(self.N_ANGLES):
+                            bdv_writer.append_view(stack=None,
+                                                   virtual_stack_dim=(self.NZ, self.NY, self.NX),
+                                                   time=t, channel=i_ch, illumination=i_illum,
+                                                   tile=i_tile, angle=i_angle)
+        # Populate the virtual stacks
+        for t in range(self.N_T):
+            for i_ch in range(self.N_CH):
+                for i_illum in range(self.N_ILL):
+                    for i_tile in range(self.N_TILES):
+                        for i_angle in range(self.N_ANGLES):
+                            for isub in range(self.N_SUBSTACKS):
+                                bdv_writer.append_substack(substack=self.substack,
+                                                           z_start=isub*self.substack.shape[0],
+                                                           time=t, channel=i_ch, illumination=i_illum,
+                                                           tile=i_tile, angle=i_angle)
+        bdv_writer.write_xml_file(ntimes=self.N_T)
+        bdv_writer.close()
+
+    def test_read_voxels(self):
+        """"Read the voxel values back and compare to expected values"""
+        editor = npy2bdv.BdvEditor(self.fname1)
+        for t in range(self.N_T):
+            for i_ch in range(self.N_CH):
+                for i_illum in range(self.N_ILL):
+                    for i_tile in range(self.N_TILES):
+                        for i_angle in range(self.N_ANGLES):
+                            view = editor.read_view(time=t, channel=i_ch, illumination=i_illum,
+                                                    tile=i_tile, angle=i_angle)
+                            self.assertAlmostEqual(np.mean(view), np.mean(self.substack), 6, "Mean differs.")
+                            self.assertAlmostEqual(np.std(view), np.std(self.substack), 6, "Std differs.")
+        editor.finalize()
+
+    def tearDown(self) -> None:
+        """"Tidies up after EACH test method has been run."""
         if os.path.exists(self.test_dir):
             shutil.rmtree(self.test_dir)