+++ CHANGELOG.md
@@ -1,5 +1,9 @@
-## v.1.0.2 candidate
+## v.1.0.3 candidate
+
 :exclamation: **API change**: The `BdvReader` class was replaced with `BdvEditor`, which allows e.g. streamlined reading and cropping views in H5 and XML files.
+
+:gem: new function `BdvWriter.append_substack()` to write substacks into a virtual stack. 
+This allows saving virtual stacks with multi-resolution pyramids downsampled in `z` as well as `(x,y)`.
 
 :gem: `BdvEditor.append_affine()`: affine transformation can be appended to an existing dataset (before, it was only possible when creating a new dataset via `BdvWriter.write_xml_file()`).